responsiveness and import/export of functions
