+++ index.html
@@ -133,7 +133,9 @@
             </div>
         </div>
        <div class="hero-video-container" >
-        
+            <div class="video-esc-box">
+
+            </div>
         </div>   
     </section>
     
@@ -858,7 +860,7 @@
     <button onclick="topFunction()"id="backToTopBtn" >
     <div class="fa fa-angle-up center"></div>
     </button>
-    <script src="./js/pages/home.js" defer></script>
+    <script src="./js/pages/home.js" type="module" defer></script>
     <script src="./js/components/backtotop.js" defer></script>
   </body>
 </html>
